Added the ability to load modules on tiles from XML, show modules on tiles to the user, and save modules on tiles to XML. Recursive modules are not yet supported; nor are tiles with multiple modules on them.

+++ sp_java_next/model/module/RootModule.java
@@ -140,7 +140,7 @@ public final class RootModule implements Module, Serializable {
 	 * @return the module's UUID
 	 */
 	@Override
-	public int getUuid() {
+	public long getUuid() {
 		return 0;
 	}
 }
